Updated Grid System
Now works with Vector positions instead of actual GameObject tiles.
- Uses less data storage
- Works with the new placement system
- Is easy scriptable to do the same on walls if we want to.
- A bit less accurate with borders etc.

+++ VR Architecture Simulation/Assets/Scripts/Options.cs
@@ -21,7 +21,7 @@ public class Options : UIMenu
 
     }
 
-    public void UpdateGridDivision(int newDivValue)
+    public void UpdateGridDivision(float newDivValue)
     {
         gridDivText.text = newDivValue.ToString();
     }
